feat: Improve the style of changing the theme
